pyproject.toml
[project]
name = "boostburn"
version = "0.1.0"
description = "LangGraph workflow for Bedrock usage reporting"
requires-python = ">=3.13"
dependencies = [
  "boto3>=1.34.0",
  "beautifulsoup4>=4.12.0",
  "langgraph>=0.2.0",
  "PyYAML>=6.0.1",
  "requests>=2.31.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0.0",
]

[project.scripts]
boostburn = "boostburn.cli:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
